Added baubles api for future use in enhancement/recharge items. Added "living essence" entities that fly around and get absorbed by nearby players' magical items, as a free but slow early-game source of elemental essences. Updated to 1.8.9 forge.

+++ src/main/java/gigaherz/elementsofpower/database/MagicAmounts.java
@@ -59,9 +59,9 @@ public class MagicAmounts
         return true;
     }
 
-    public int getTotalMagic()
+    public float getTotalMagic()
     {
-        int acc = 0;
+        float acc = 0;
 
         for (float amount : amounts)
         {
@@ -156,7 +156,8 @@ public class MagicAmounts
     }
 
     public static class Serializer
-            implements JsonSerializer<MagicAmounts>, JsonDeserializer<MagicAmounts>
+            implements JsonSerializer<MagicAmounts>,
+            JsonDeserializer<MagicAmounts>
     {
         @Override
         public JsonElement serialize(MagicAmounts src, Type typeOfSrc, JsonSerializationContext context)
